requirements.txt
boto3
faker
kafka-python
